requirements.txt
streamlit==1.35.0
web3==6.18.0
pandas==2.1.4
python-dotenv==1.0.1
matplotlib==3.8.4
